// backend/src/types.ts
export interface MeritTopper {
  rank: number;
  registration_id: string;
  roll_no: string;
  name: string;
  father_name: string;
  school_name: string;
  district: string;
  current_class: string;
  category: string;
  marks_obtained: number;
  total_marks: number;
  percentage: number;
  award_scholarship: string;
  photo_url?: string;
}